dbms/glau/bcsc0034/lecture1/lab_cheatsheet/requirements.txt
# DBMS Lab Dependencies
mysql-connector-python
